implements parameter class
removes the parameter struct from the Struct file and turn it into a
class.
